+++ src/Cwaff-Misc/CwaffSynergies.cs
@@ -52,7 +52,7 @@ public static class CwaffSynergies
         NewSynergy(SOUL_SEARCHING, "Soul Searching", new[]{IName(Uppskeruvel.ItemName), "gun_soul"});
         // Vacuuming debris occasionally generates casings (up to 20 per floor).
         NewSynergy(CLEANUP_CREW, "Cleanup Crew", new[]{IName(VacuumCleaner.ItemName), IName(CustodiansBadge.ItemName)});
-        // Vacuum's chance to restore ammo is increased to 4%.
+        // Vacuum's chance to restore ammo is increased to 20%.
         NewSynergy(SCAVENGEST, "Scavengest", new[]{IName(VacuumCleaner.ItemName), IName(ScavengingArms.ItemName)});
         // When taking an otherwise fatal hit, if the player has at least 100 casings, damage is negated and the player loses 100 casings instead.
         NewSynergy(DEATH_AND_TAXES, "Death and Taxes", new[]{IName(CreditCard.ItemName), IName(BlankChecks.ItemName)});